// src/utils/enums/ChartEnum.ts
export enum BarChartType {
    NORMAL = "normal",
    GROUPED = "grouped",
    STACKED = "stacked",
}

export enum RadialBarChartType {
    NORMAL = "normal",
    INSIGHT = "insight",
    COMPACT = "compact"
}

export enum LegendAlignType {
    BOTTOM = "bottom",
    RIGHT = "right"
}

export enum LegendType {
    A = "A",
    B = "B",
    C = "C",
    D = "D",
    E = "E",
    F = "F",
    AA = "AA",
    BB = "BB",
    CC = "CC",
    DD = "DD",
    EE = "EE",
    FF = "FF",
}